[trigger_handler] Enhance _findTriggerConfig
There are two events.
- The one is the event from the configuration that the user specify.
This event can be made with main event and sub event both or main
event only.
- The other is the event from the repository. This event can be made
only with main event and sub event both.

If two events are match, a task will be triggered.
If the first event is made with main event only and the second event
is made with same main event and whatever sub event, a task will be
triggered.